fix(repositories): preserve imports across connections
Keep provider identities imported across all matching connections without reassigning catalog ownership, report only newly created rows, renumber the migration after rebasing on master, and consolidate provider test scaffolding to stay within the PR file limit.
