--- server/src/services/repository-providers/index.ts
@@ -1,10 +1,0 @@
-export * from "./provider-contract.js";
-export * from "./connection-state.js";
-export * from "./github-provider.js";
-export * from "./github-transport.js";
-export * from "./fake-provider.js";
-export * from "./registry.js";
-export * from "./connector-guard.js";
-export * from "./plugin-connector.js";
-export * from "./plugin-registrar.js";
-export * from "./bootstrap.js";
